+++ vite.config.ts
@@ -7,12 +7,21 @@
 
 import { resolve } from "node:path";
 import { defineConfig } from "vite";
+import dts from "vite-plugin-dts";
+import tsconfigPaths from "vite-tsconfig-paths";
 import react from "@vitejs/plugin-react-swc";
 import { libInjectCss } from "vite-plugin-lib-inject-css";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-	plugins: [react(), libInjectCss()],
+	plugins: [
+		react(),
+		dts({
+			insertTypesEntry: true,
+		}),
+		tsconfigPaths(),
+		libInjectCss(),
+	],
 	build: {
 		lib: {
 			entry: resolve(__dirname, "src/index.js"),
@@ -24,7 +33,7 @@ export default defineConfig({
 					umd: "index.umd.cjs",
 				};
 
-				return OUTPUT[format] ?? "index.js";
+				return OUTPUT[format] ?? "index.umd.cjs";
 			},
 		},
 		rollupOptions: {
@@ -34,6 +43,7 @@ export default defineConfig({
 					react: "React",
 					"react-dom": "ReactDOM",
 					"react/jsx-runtime": "react/jsx-runtime",
+					"@feedzai/js-utilities": "JSUtilities",
 				},
 			},
 		},