+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2.0.3 - 2024-11-04
+### Added
+- Adds missing types
+- Fixed the issue where components were not being properly exported
+
 ## 2.0.2 - 2024-10-24
 ### Added
 - Dynamic changes in the data or DOM elements now trigger updates in the tool